+++ app/Controller/Component/ApiComponent.php
@@ -225,7 +225,20 @@ class ApiComponent extends Component {
           }
         }
       }
-      
+
+      if ($this->reqModelName == 'Cou'
+          && !empty($rr['Cou'])
+          && !empty($res[0]['CoDepartment'])) {
+        foreach ($res[0]['CoDepartment'] as $idx => $department) {
+          $rr['Cou']['Metadata'][] = array();
+          foreach ($department as $key => $value) {
+            if(in_array($key, array('type'))) {
+              $rr['Cou']['Metadata'][$idx][Inflector::camelize($key)] = $value;
+            }
+          }
+        }
+      }
+
       $ret[] = $rr;
     }