+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [] - xxx-xx-xx
 
+### Changed
+
+- COU REST API will include the department type
+
 ## [3.4.6-rciam] - 2022-05-17
 
 ### Added